+++ src/Symfony/Component/Notifier/Bridge/Telegram/TelegramTransportFactory.php
@@ -31,9 +31,9 @@ final class TelegramTransportFactory extends AbstractTransportFactory
         $channel = $dsn->getOption('channel');
         $host = 'default' === $dsn->getHost() ? null : $dsn->getHost();
         $port = $dsn->getPort();
-        $useHttpProtocol = '1' === $dsn->getOption('use_http_protocol');
+        $disableHttps = $this->isHttpsDisabled($dsn);
 
-        return (new TelegramTransport($token, $channel, $useHttpProtocol, $this->client, $this->dispatcher))->setHost($host)->setPort($port);
+        return (new TelegramTransport($token, $channel, $this->client, $this->dispatcher, $disableHttps))->setHost($host)->setPort($port);
     }
 
     protected function getSupportedSchemes(): array
@@ -52,5 +52,18 @@ final class TelegramTransportFactory extends AbstractTransportFactory
         }
 
         return \sprintf('%s:%s', $dsn->getUser(), $dsn->getPassword());
+    }
+
+    private function isHttpsDisabled(Dsn $dsn): bool
+    {
+        if (!$disableHttpsOption = $dsn->getOption('disable_https')) {
+            return false;
+        }
+
+        if (!\in_array($disableHttpsOption, [0, 1])) {
+            throw new IncompleteDsnException(\sprintf('Only 1 or 0 are allowed as a value for option "disable_https", "%s" given.', $disableHttpsOption), \sprintf('telegram://%s', $dsn->getHost()));
+        }
+
+        return (bool) $disableHttpsOption;
     }
 }